refactor(schedule): split DroppableTrack/ScheduleEditor into focused files
Phase 4-5 of the schedule-editor restructure. Behaviour-preserving except two
enumerated fixes.

- Split the ~1026-line DroppableTrack god-component into a track/ folder: one
  file each for TrackHeader, ScheduledTalk, ServiceSession, ServiceSessionModal
  and TimeSlotDropZone; DroppableTrack becomes a thin composing container.
- Extract AddTrackModal and HeaderSection out of ScheduleEditor.
- Add src/lib/schedule/geometry.ts as the single source of PIXELS_PER_MINUTE and
  the pixel position helpers; drop the redeclared MINUTES_TO_PIXELS/PIXELS_PER_MINUTE
  in DroppableTrack, DraggableProposal and DraggableServiceSession.
- Add ScheduleContext to provide activeDragItem, the current schedule and
  dispatch to the leaf drop targets, killing the activeDragItem prop-drilling
  through TracksGrid -> DroppableTrack.

FIX #1: extract the service-session resize into useServiceSessionResize and
convert it from mouse-only document listeners to Pointer Events
(onPointerDown + pointermove/pointerup with setPointerCapture) so resizing works
under touch. The new end is clamped to stay free (rules.isTrackIntervalFree) and
within SCHEDULE_END.

FIX #2: TimeSlotDropZone.canDrop now also runs the reverse swap check
(rules.canPlaceDisplacedBack) using the schedule from context, so the drop
indicator matches what the reducer actually does instead of showing swaps as
droppable that the drop then rejects.
